create test-common module and move some util out of the helper-specific test module
