+++ helper/tests/codegen-16/src/main/java/com/linkedin/avroutil1/compatibility/avro16/Generate16TestResources.java
@@ -6,7 +6,7 @@
 
 package com.linkedin.avroutil1.compatibility.avro16;
 
-import com.linkedin.avroutil1.TestUtil;
+import com.linkedin.avroutil1.testcommon.TestUtil;
 import org.apache.avro.io.BinaryEncoder;
 import org.apache.avro.io.EncoderFactory;
 import org.apache.avro.io.JsonEncoder;
