+++ helper/tests/helper-tests-14/src/test/java/com/linkedin/avroutil1/compatibility/avro14/Avro14GeneratedCodeTest.java
@@ -6,7 +6,7 @@
 
 package com.linkedin.avroutil1.compatibility.avro14;
 
-import com.linkedin.avroutil1.TestUtil;
+import com.linkedin.avroutil1.testcommon.TestUtil;
 import com.linkedin.avroutil1.compatibility.AvroCompatibilityHelper;
 import net.openhft.compiler.CompilerUtils;
 import org.apache.avro.Schema;
