+++ helper/tests/helper-tests-15/src/test/java/com/linkedin/avroutil1/compatibility/avro15/CodeTransformationsAvro15Test.java
@@ -6,7 +6,7 @@
 
 package com.linkedin.avroutil1.compatibility.avro15;
 
-import com.linkedin.avroutil1.TestUtil;
+import com.linkedin.avroutil1.testcommon.TestUtil;
 import com.linkedin.avroutil1.compatibility.AvroCompatibilityHelper;
 import com.linkedin.avroutil1.compatibility.AvroVersion;
 import com.linkedin.avroutil1.compatibility.CodeTransformations;
