+++ helper/tests/helper-tests-16/src/test/java/com/linkedin/avroutil1/compatibility/avro16/Avro16ParseBehaviorTest.java
@@ -6,7 +6,7 @@
 
 package com.linkedin.avroutil1.compatibility.avro16;
 
-import com.linkedin.avroutil1.TestUtil;
+import com.linkedin.avroutil1.testcommon.TestUtil;
 import org.apache.avro.Schema;
 import org.apache.avro.SchemaParseException;
 import org.testng.Assert;
