+++ helper/tests/helper-tests-16/src/test/java/com/linkedin/avroutil1/compatibility/avro16/CodeTransformationsAvro16Test.java
@@ -6,7 +6,7 @@
 
 package com.linkedin.avroutil1.compatibility.avro16;
 
-import com.linkedin.avroutil1.TestUtil;
+import com.linkedin.avroutil1.testcommon.TestUtil;
 import com.linkedin.avroutil1.compatibility.AvroCompatibilityHelper;
 import com.linkedin.avroutil1.compatibility.AvroVersion;
 import com.linkedin.avroutil1.compatibility.CodeTransformations;
