+++ helper/tests/helper-tests-allavro/src/test/java/com/linkedin/avroutil1/compatibility/SchemaBuilderTest.java
@@ -6,7 +6,7 @@
 
 package com.linkedin.avroutil1.compatibility;
 
-import com.linkedin.avroutil1.TestUtil;
+import com.linkedin.avroutil1.testcommon.TestUtil;
 import org.apache.avro.Schema;
 import org.testng.Assert;
 import org.testng.annotations.Test;
