+++ settings.gradle
@@ -20,6 +20,7 @@ apply plugin: 'net.vivin.gradle-semantic-build-versioning'
 //otherwise it defaults to the folder name
 rootProject.name = 'avro-util'
 
+include 'test-common'
 include 'avro-codegen'
 include 'avro-fastserde'
