+++ spotbugs-plugin/build.gradle
@@ -13,6 +13,7 @@ plugins {
 dependencies {
   compileOnly "com.github.spotbugs:spotbugs:4.2.3"
 
+  testCompile project(":test-common")
   testCompile project(":helper:tests:helper-tests-common")
 
   testImplementation project(":helper:helper")
