+++ spotbugs-plugin/src/test/java/com/linkedin/avroutil1/spotbugs/BadClass.java
@@ -6,7 +6,7 @@
 
 package com.linkedin.avroutil1.spotbugs;
 
-import com.linkedin.avroutil1.TestUtil;
+import com.linkedin.avroutil1.testcommon.TestUtil;
 import org.apache.avro.Schema;
 import org.apache.avro.generic.GenericData;
 import org.apache.avro.generic.GenericRecord;
