+++ spotbugs-plugin/src/test/java/com/linkedin/avroutil1/spotbugs/OldSchemaConstructableUsageDetectorTest.java
@@ -6,7 +6,7 @@
 
 package com.linkedin.avroutil1.spotbugs;
 
-import com.linkedin.avroutil1.TestUtil;
+import com.linkedin.avroutil1.testcommon.TestUtil;
 import edu.umd.cs.findbugs.BugCollection;
 import edu.umd.cs.findbugs.test.SpotBugsRule;
 import edu.umd.cs.findbugs.test.matcher.BugInstanceMatcher;
