+++ test-common/build.gradle
@@ -1,0 +1,16 @@
+/*
+ * Copyright 2021 LinkedIn Corp.
+ * Licensed under the BSD 2-Clause License (the "License").
+ * See License in the project root for license information.
+ */
+
+plugins {
+  id "java-library"
+  id "checkstyle"
+}
+
+dependencies {
+  //we use this module as an easy way to "export" libraries for use by other test modules
+  api "commons-io:commons-io:2.6"
+  api "net.openhft:compiler:2.3.6"
+}
